backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-dotenv>=1.0.0
anthropic>=0.40.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
osmnx>=1.9.0
networkx>=3.2
geopandas>=0.14.0
shapely>=2.0.0
numpy>=1.26.0
httpx>=0.27.0
aiofiles>=24.0.0
pydantic>=2.7.0
sse-starlette>=2.1.0
